AR_Design-main/requirements.txt
flask==2.0.1
pillow==9.0.0
numpy==1.21.0
requests==2.26.0
python-dotenv==0.19.0
opencv-python==4.5.3.56
opencv-contrib-python==4.5.3.56 